+++ packages/flint-js/tsup.config.ts
@@ -7,7 +7,6 @@ export default defineConfig({
     'vegalite/index': 'src/vegalite/index.ts',
     'echarts/index': 'src/echarts/index.ts',
     'chartjs/index': 'src/chartjs/index.ts',
-    'gofish/index': 'src/gofish/index.ts',
     'test-data/index': 'src/test-data/index.ts',
     'gallery/index': 'src/gallery/index.ts',
   },
@@ -18,5 +17,5 @@ export default defineConfig({
   splitting: false,
   treeshake: true,
   target: 'es2020',
-  external: ['vega', 'vega-lite', 'echarts', 'chart.js', 'gofish-graphics'],
+  external: ['vega', 'vega-lite', 'echarts', 'chart.js'],
 });